Now for the *important* bits :-D
Every value that comes from the form and goes into the SQL string is _escaped_. That makes sure characters like single quotes are
  prepended with a slash, so they don't muck up the resulting query (you may have noticed checking method "TEN's pads" broke the
  insert).
  Keep in mind that SQL is a programming language of its own, and what PHP is doing here is essentially "writing code in a different
  language", which is then run by MySQL. Malicious users can enter specially crafted SQL code in the inputs, which then becomes part
  of the query, and which is then run in the database. An insert will not echo query results back to the page, but if it was a
  search they could add UNION statements on user tables to yield password hashes for example. As is with the insert, one could add
  or drop tables.. or add users. It depends on what permissions the database user has that was used to create the connection, but
  it's a security issue.
  It's also, year after year, the #1 way that websites get compromised: https://owasp.org/www-project-top-ten/

Minor stuff: changed name 'method' to 'methods' to reflect the fact that you can choose several. Really you want all names to be
  accurate, and to be the same in the form, code, and database, this prevents a lot of simple errors.
Added labels around checkbox inputs and the text so you can click on the text and it will toggle the checkbox.
Tweaked CSS a little so the top div grows in height as needed; submitted notes now show with linebreaks, and choosing many methods
  or getting SQL errors can also add text.